templates: fix published status
The published status was being reported as `bool(build.published)`,
however BuildRecord has no attribute `published`. Therefore Django's
template engine defaults it to`None` which evaluates to `False`.

This change adds the template context variable `published` which is
evaluated to `publisher.published(build)` and that variable is instead
used for the published status.